finish handle seller product in admin

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -18,8 +19,10 @@ class VendorProductImageGalleryController extends Controller
     public function index(Request $request, VendorProductImageGalleryDataTable $dataTable)
     {
         //
-        $product = Product::find($request->product);
-
+        $product = Product::findOrFail($request->product);
+        if ($product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
         return  $dataTable->render("vendor.products.image-gallery.index", compact('product'));
     }
 
@@ -82,6 +85,9 @@ class VendorProductImageGalleryController extends Controller
     {
         //
         $productImageGallery = ProductImageGallery::find($id);
+        if ($productImageGallery->product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
         $this->deleteImage($productImageGallery->image);
         $productImageGallery->delete();
         return response([
